Improve performance testing: retry on failure due to machine load.
