+++ test/list/list_builder_test.dart
@@ -9,6 +9,8 @@ import 'dart:math' show Random;
 import 'package:built_collection/built_collection.dart';
 import 'package:test/test.dart';
 
+import '../performance.dart';
+
 void main() {
   group('ListBuilder', () {
     test('throws on attempt to create ListBuilder<dynamic>', () {
@@ -364,23 +366,6 @@ void main() {
           [2]);
     });
   });
-}
-
-void expectMuchFaster(Function fastFunction, Function slowFunction) {
-  final fastStopWatch = new Stopwatch()..start();
-  fastFunction();
-  fastStopWatch.stop();
-
-  final slowStopWatch = new Stopwatch()..start();
-  slowFunction();
-  slowStopWatch.stop();
-
-  if (fastStopWatch.elapsedMicroseconds * 10 >
-      slowStopWatch.elapsedMicroseconds) {
-    throw 'Expected first function to be at least 10x faster than second!'
-        ' Measured: first=${fastStopWatch.elapsedMicroseconds}'
-        ' second=${slowStopWatch.elapsedMicroseconds}';
-  }
 }
 
 class _AlwaysZeroRandom implements Random {
