+++ test/list_multimap/built_list_multimap_test.dart
@@ -9,6 +9,8 @@ import 'package:built_collection/src/internal/test_helpers.dart';
 import 'package:quiver/collection.dart';
 import 'package:test/test.dart';
 
+import '../performance.dart';
+
 void main() {
   group('BuiltListMultimap', () {
     test('instantiates empty by default', () {
@@ -600,40 +602,6 @@ void main() {
       expect(multimap.values, same(multimap.values));
     });
   });
-}
-
-void expectMuchFaster(Function fastFunction, Function slowFunction) {
-  final fastStopWatch = new Stopwatch()..start();
-  fastFunction();
-  fastStopWatch.stop();
-
-  final slowStopWatch = new Stopwatch()..start();
-  slowFunction();
-  slowStopWatch.stop();
-
-  if (fastStopWatch.elapsedMicroseconds * 10 >
-      slowStopWatch.elapsedMicroseconds) {
-    throw 'Expected first function to be at least 10x faster than second!'
-        ' Measured: first=${fastStopWatch.elapsedMicroseconds}'
-        ' second=${slowStopWatch.elapsedMicroseconds}';
-  }
-}
-
-void expectNotMuchFaster(Function notFastFunction, Function slowFunction) {
-  final fastStopWatch = new Stopwatch()..start();
-  notFastFunction();
-  fastStopWatch.stop();
-
-  final slowStopWatch = new Stopwatch()..start();
-  slowFunction();
-  slowStopWatch.stop();
-
-  if (fastStopWatch.elapsedMicroseconds * 10 <
-      slowStopWatch.elapsedMicroseconds) {
-    throw 'Expected first function to be less than 10x faster than second!'
-        ' Measured: first=${fastStopWatch.elapsedMicroseconds}'
-        ' second=${slowStopWatch.elapsedMicroseconds}';
-  }
 }
 
 class _A {}
