+++ test/map/map_builder_test.dart
@@ -7,6 +7,8 @@ library built_collection.test.map.map_builder_test;
 import 'package:built_collection/built_collection.dart';
 import 'package:test/test.dart';
 
+import '../performance.dart';
+
 void main() {
   group('MapBuilder', () {
     test('throws on attempt to create MapBuilder<dynamic, dynamic>', () {
@@ -187,21 +189,4 @@ void main() {
           {});
     });
   });
-}
-
-void expectMuchFaster(Function fastFunction, Function slowFunction) {
-  final fastStopWatch = new Stopwatch()..start();
-  fastFunction();
-  fastStopWatch.stop();
-
-  final slowStopWatch = new Stopwatch()..start();
-  slowFunction();
-  slowStopWatch.stop();
-
-  if (fastStopWatch.elapsedMicroseconds * 10 >
-      slowStopWatch.elapsedMicroseconds) {
-    throw 'Expected first function to be at least 10x faster than second!'
-        ' Measured: first=${fastStopWatch.elapsedMicroseconds}'
-        ' second=${slowStopWatch.elapsedMicroseconds}';
-  }
 }
