+++ test/set_multimap/set_multimap_builder_test.dart
@@ -8,6 +8,8 @@ import 'package:built_collection/built_collection.dart';
 import 'package:quiver/collection.dart';
 import 'package:test/test.dart';
 
+import '../performance.dart';
+
 void main() {
   group('SetMultimapBuilder', () {
     test('throws on attempt to create SetMultimapBuilder<dynamic, dynamic>',
@@ -414,21 +416,4 @@ void main() {
           {});
     });
   });
-}
-
-void expectMuchFaster(Function fastFunction, Function slowFunction) {
-  final fastStopWatch = new Stopwatch()..start();
-  fastFunction();
-  fastStopWatch.stop();
-
-  final slowStopWatch = new Stopwatch()..start();
-  slowFunction();
-  slowStopWatch.stop();
-
-  if (fastStopWatch.elapsedMicroseconds * 10 >
-      slowStopWatch.elapsedMicroseconds) {
-    throw 'Expected first function to be at least 10x faster than second!'
-        ' Measured: first=${fastStopWatch.elapsedMicroseconds}'
-        ' second=${slowStopWatch.elapsedMicroseconds}';
-  }
 }
